+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improve export exception handler
 - Remove jvm started flaky metric
+- Fix `event/op-completed` metric to measure time correctly
+- Add `event/test-executed` event.
 
 ## 0.1.2